+++ src/Custom-Hooks/useTheme.tsx
@@ -1,0 +1,22 @@
+import { useEffect } from "react";
+import useLocalStorage from "use-local-storage";
+
+const useTheme = () => {
+  const defaultDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
+  const [theme, setTheme] = useLocalStorage('theme', defaultDark ? 'dark' : 'light');
+
+  const toggleTheme = () => {
+    setTheme(theme === 'light' ? 'dark' : 'light');
+  }
+
+  useEffect(() => {      
+    document.body.setAttribute('data-theme', theme);
+    return () => {
+      document.body.removeAttribute('data-theme');
+    };
+  }, [theme]);
+
+  return [theme, toggleTheme];
+}
+
+export default useTheme;
